refactor(test-optimization): keep request agents isolated
